source/mre-video-ingestion-samples/HLS_Harvester/infrastructure/requirements.txt
aws_cdk.core>=1.85.0,<2.0
aws_cdk.aws_iam>=1.85.0,<2.0
chalice[cdk]
